+++ src/DataFilters/Grammar/Syntax/ContainsExpression.cs
@@ -5,7 +5,11 @@ namespace DataFilters.Grammar.Syntax
     /// <summary>
     /// A <see cref="FilterExpression"/> that holds a string value
     /// </summary>
+#if NETSTANDARD1_3
     public sealed class ContainsExpression : FilterExpression, IEquatable<ContainsExpression>
+#else
+    public record ContainsExpression : FilterExpression, IEquatable<ContainsExpression>
+#endif
     {
         /// <summary>
         /// The value that was between two <see cref="AsteriskExpression"/>
@@ -27,19 +31,28 @@ namespace DataFilters.Grammar.Syntax
 
             if (value.Length == 0)
             {
-                throw new ArgumentOutOfRangeException(nameof(value));
+                throw new ArgumentOutOfRangeException(nameof(value), $"{nameof(value)} cannot be empty");
             }
 
             Value = value;
         }
 
         ///<inheritdoc/>
-        public bool Equals(ContainsExpression other) => Value == other?.Value;
+        public override double Complexity => 1.5;
 
+#if NETSTANDARD1_3
         ///<inheritdoc/>
         public override bool Equals(object obj) => Equals(obj as ContainsExpression);
 
         ///<inheritdoc/>
+        public bool Equals(ContainsExpression other) => Value == other?.Value;
+
+        ///<inheritdoc/>
         public override int GetHashCode() => Value.GetHashCode();
+
+        ///<inheritdoc/>
+        public override string ToString() => this.Jsonify();
+
+#endif
     }
 }