+++ src/DataFilters/Grammar/Syntax/GroupExpression.cs
@@ -3,9 +3,21 @@
 namespace DataFilters.Grammar.Syntax
 {
     /// <summary>
-    /// An expression that group wrapped inside
+    /// Allows to treat several expressions as a single unit.
     /// </summary>
+    /// <remarks>
+    /// <para>
+    /// A group expression can be used whenever there is a need to apply a logical operator to several expressions at once.
+    /// </para>
+    /// <para>
+    /// The <see cref="Complexity"/> value of a <see cref="GroupExpression"/> is equivalent to the complexity of its inner <see cref="Expression"/>.
+    /// </para>
+    /// </remarks>
+#if NETSTANDARD1_3
     public sealed class GroupExpression : FilterExpression, IEquatable<GroupExpression>
+#else
+    public record GroupExpression : FilterExpression, IEquatable<GroupExpression>
+#endif
     {
         /// <summary>
         /// Expression that the group is applied onto
@@ -13,12 +25,13 @@ namespace DataFilters.Grammar.Syntax
         public FilterExpression Expression { get; }
 
         /// <summary>
-        /// Builds a new <see cref="GroupExpression"/> that holds the specified <paramref name="innerExpression"/>.
+        /// Builds a new <see cref="GroupExpression"/> that holds the specified <paramref name="expression"/>.
         /// </summary>
-        /// <param name="innerExpression"></param>
-        /// <exception cref="ArgumentNullException"><paramref name="innerExpression"/> is <c>null</c>.</exception>
-        public GroupExpression(FilterExpression innerExpression) => Expression = innerExpression ?? throw new ArgumentNullException(nameof(innerExpression));
+        /// <param name="expression"></param>
+        /// <exception cref="ArgumentNullException"><paramref name="expression"/> is <c>null</c>.</exception>
+        public GroupExpression(FilterExpression expression) => Expression = expression ?? throw new ArgumentNullException(nameof(expression));
 
+#if NETSTANDARD1_3
         ///<inheritdoc/>
         public bool Equals(GroupExpression other) => Expression.Equals(other?.Expression);
 
@@ -30,5 +43,12 @@ namespace DataFilters.Grammar.Syntax
 
         ///<inheritdoc/>
         public override string ToString() => $"{GetType().Name} : Expression ({Expression.GetType().Name }) -> {Expression}";
+#endif
+
+        ///<inheritdoc/>
+        public override double Complexity => Expression.Complexity;
+
+        ///<inheritdoc/>
+        public override bool IsEquivalentTo(FilterExpression other) => Expression.IsEquivalentTo(other);
     }
 }