+++ src/DataFilters/Grammar/Syntax/NotExpression.cs
@@ -5,7 +5,11 @@ namespace DataFilters.Grammar.Syntax
     /// <summary>
     /// An expression that negate wrapped inside
     /// </summary>
+#if NETSTANDARD1_3
     public sealed class NotExpression : FilterExpression, IEquatable<NotExpression>
+#else
+    public record NotExpression : FilterExpression, IEquatable<NotExpression>
+#endif
     {
         /// <summary>
         /// Expression that the NOT logical is applied to
@@ -13,12 +17,13 @@ namespace DataFilters.Grammar.Syntax
         public FilterExpression Expression { get; }
 
         /// <summary>
-        /// Builds a new <see cref="NotExpression"/> that holds the specified <paramref name="innerExpression"/>.
+        /// Builds a new <see cref="NotExpression"/> that holds the specified <paramref name="expression"/>.
         /// </summary>
-        /// <param name="innerExpression"></param>
-        /// <exception cref="ArgumentNullException"><paramref name="innerExpression"/> is <c>null</c>.</exception>
-        public NotExpression(FilterExpression innerExpression) => Expression = innerExpression ?? throw new ArgumentNullException(nameof(innerExpression));
+        /// <param name="expression"></param>
+        /// <exception cref="ArgumentNullException"><paramref name="expression"/> is <c>null</c>.</exception>
+        public NotExpression(FilterExpression expression) => Expression = expression ?? throw new ArgumentNullException(nameof(expression));
 
+#if NETSTANDARD1_3
         ///<inheritdoc/>
         public bool Equals(NotExpression other) => Expression.Equals(other?.Expression);
 
@@ -30,5 +35,9 @@ namespace DataFilters.Grammar.Syntax
 
         ///<inheritdoc/>
         public override string ToString() => $"{GetType().Name} : Expression ({Expression.GetType().Name}) -> {Expression}";
+#endif
+
+        ///<inheritdoc/>
+        public override double Complexity => Expression.Complexity;
     }
 }