+++ src/DataFilters/Grammar/Syntax/OrExpression.cs
@@ -5,7 +5,11 @@ namespace DataFilters.Grammar.Syntax
     /// <summary>
     /// An expression that combine two <see cref="FilterExpression"/> expressions using the logical <c>OR</c> operator
     /// </summary>
+#if NETSTANDARD1_3
     public sealed class OrExpression : FilterExpression, IEquatable<OrExpression>
+#else
+    public record OrExpression : FilterExpression, IEquatable<OrExpression>
+#endif
     {
         /// <summary>
         /// Left member of the expression
@@ -30,6 +34,7 @@ namespace DataFilters.Grammar.Syntax
             Right = right ?? throw new ArgumentNullException(nameof(right));
         }
 
+#if NETSTANDARD1_3
         ///<inheritdoc/>
         public bool Equals(OrExpression other) => Left.Equals(other?.Left) && Right.Equals(other?.Right);
 
@@ -50,22 +55,27 @@ namespace DataFilters.Grammar.Syntax
             Right = new { Type = Right.GetType().Name, Right },
         }.Jsonify();
 
+#endif
         /// <inheritdoc/>
         public override bool IsEquivalentTo(FilterExpression other)
         {
             bool equivalent = false;
 
-            if (other is OrExpression or)
+            switch (other)
             {
-                equivalent = or.Right.Equals(Right) && or.Left.Equals(Left)
-                    || (or.Left.Equals(Right) && or.Right.Equals(Left));
-            }
-            else
-            {
-                equivalent = Left.Equals(Right) && Left.Equals(other);
+                case OrExpression or:
+                    equivalent = (or.Right.Equals(Right) && or.Left.Equals(Left))
+                                 || (or.Left.Equals(Right) && or.Right.Equals(Left));
+                    break;
+                default:
+                    equivalent = Left.Equals(Right) && Left.Equals(other);
+                    break;
             }
 
             return equivalent;
         }
+
+        ///<inheritdoc/>
+        public override double Complexity => Left.Complexity + Right.Complexity;
     }
 }